+++ paypal/models.py
@@ -1,3 +1,21 @@
-from django.db import models
+import enum
+from dataclasses import dataclass
 
-# Create your models here.
+
+class PayPalError(Exception):
+    pass
+
+
+class CurrencyCode(enum.Enum):
+    USD = "USD"
+
+
+@dataclass
+class PayPalAmount:
+    currency_code: CurrencyCode
+    value: str
+
+
+@dataclass
+class PayPalPurchaseUnit:
+    amount: PayPalAmount
