+++ paypal/subscriptions.py
@@ -1,8 +1,13 @@
+import logging
 from django.core.cache import cache
 import requests
+from requests.exceptions import HTTPError
 
+from .auth import construct_paypal_auth_headers
 from paypal.constants import ONE_DAY_S, PAYPAL_SUBSCRIPTIONS_BASE_URL
-from paypal.utils import construct_paypal_auth_headers
+from paypal.models import PayPalError
+
+logger = logging.getLogger(__name__)
 
 def get_subscription(
     *,
@@ -14,6 +19,12 @@ def get_subscription(
         url=f"{PAYPAL_SUBSCRIPTIONS_BASE_URL}/{paypal_subscription_id}",
         headers=headers,
     )
+
+    try:
+        response.raise_for_status()
+    except HTTPError as error:
+        logger.exception(error)
+        raise PayPalError(error)
 
     return response.json()
 
@@ -28,9 +39,12 @@ def subscription_is_active(
     status = cache.get(cache_key)
 
     if status is None:
-        subscription = get_subscription(
-            paypal_subscription_id=paypal_subscription_id,
-        )
+        try:
+            subscription = get_subscription(
+                paypal_subscription_id=paypal_subscription_id,
+            )
+        except PayPalError:
+            return False
 
         status = subscription["status"]
